Unifiy Formhocs and move structure folders where they should be from a Domain Driven design perspective
