--- frontend/src/app/components/Pages/ErrorBoundary/index.js
@@ -1,1 +1,0 @@
-export { default } from './ErrorBoundary';
