+++ frontend/src/app/index.jsx
@@ -3,16 +3,16 @@ import { render } from 'react-dom';
 import { Provider } from 'react-redux';
 import { store } from './store';
 import Main from './components/Main';
-import ErrorBoundary from './components/Pages/ErrorBoundary';
+import Error from './components/Pages/Error';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'react-bootstrap-table-next/dist/react-bootstrap-table2.min.css';
 import './assets/styles/global.scss';
 
 render(
-  <ErrorBoundary>
+  <Error>
     <Provider store={store}>
       <Main />
     </Provider>
-  </ErrorBoundary>,
+  </Error>,
   document.getElementById('app'),
 );
